change mp_digit/mp_word to unsigned long long
x32 target is 64 bit arch but has 32 bit wide unsigned long

+++ tommath.h
@@ -73,8 +73,8 @@ extern "C" {
    typedef signed long long   long64;
 #endif
 
-   typedef unsigned long      mp_digit;
-   typedef unsigned long      mp_word __attribute__ ((mode(TI)));
+   typedef unsigned long long mp_digit;
+   typedef unsigned long long mp_word __attribute__ ((mode(TI)));
 
    #define DIGIT_BIT          60
 #else
